[reorg] Move ReverseChainSingleSubst to new layout.
https://github.com/harfbuzz/harfbuzz/commit/90af2143d
