+++ src/hb/ot/layout/GSUB/mod.rs
@@ -4,5 +4,6 @@ mod ligature;
 mod ligature_set;
 mod ligature_subst;
 mod multi_subst;
+mod reverse_chain_single_subst;
 mod sequence;
 mod single_subst;
